chore: fixing test build and unit test errors
