+++ src/storageHandlers/__tests__/fileShareHandler.spec.ts
@@ -1,6 +1,7 @@
 import * as networkDrive from 'windows-network-drive'
 import { FileShareHandler } from '../../storageHandlers/fileShareHandler'
 import { StorageType } from '../../api'
+import * as winston from 'winston'
 
 jest.mock('windows-network-drive')
 
@@ -17,9 +18,10 @@ describe('FileShareHandler', () => {
 			},
 			options: {
 				basePath: '\\\\STORAGE\\public',
-				mappedNetworkedDriveTarget: 'U'
+				mappedNetworkedDriveTarget: 'U',
+				onlySelectedFiles: false // to make test use '.' and cause watcher to enter ready state
 			}
-		})
+		}, new winston.Logger({ transports: [ new winston.transports.Console() ]}))
 		try {
 			fsh0.on('error', err => fail(err))
 			await fsh0.init()
@@ -30,6 +32,10 @@ describe('FileShareHandler', () => {
 	})
 
 	it('mounts the network drive automatically', () => {
-		expect(networkDrive.mount).toBeCalled()
+		expect(networkDrive.mount).toHaveBeenCalled()
+	})
+
+	afterAll(() => {
+		fsh0.destroy()
 	})
 })
