+++ src/storageHandlers/__tests__/storageHandler.spec.ts
@@ -1,8 +1,9 @@
-import { buildStorageHandler } from '../../storageHandlers/storageHandler'
+import { buildStorageHandler } from '../../storageHandlers/storageHandlerFactory'
 import { StorageType, LocalFolderStorage, FileShareStorage } from '../../api'
 import { literal } from '../../lib/lib'
 import { LocalFolderHandler } from '../localFolderHandler'
 import { FileShareHandler } from '../fileShareHandler'
+import * as winston from 'winston'
 
 describe('buildStorageHandler', () => {
 	it('returns a new instance of a StorageHandler, based on the config', () => {
@@ -17,7 +18,8 @@ describe('buildStorageHandler', () => {
 				options: {
 					basePath: './'
 				}
-			})
+			}),
+			new winston.Logger({ transports: [ new winston.transports.Console() ]})
 		)
 
 		expect(localHandler).toBeInstanceOf(LocalFolderHandler)
@@ -34,7 +36,8 @@ describe('buildStorageHandler', () => {
 					basePath: '\\\\SERVER\\share',
 					mappedNetworkedDriveTarget: 'X'
 				}
-			})
+			}),
+			new winston.Logger({ transports: [ new winston.transports.Console() ]})
 		)
 
 		expect(shareHandler).toBeInstanceOf(FileShareHandler)
